requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
pyinstaller>=5.0
